Refractored the backend in modular-MVC achitecture , implemented Outh in structured way
