+++ backend/controllers/authController.js
@@ -1,0 +1,144 @@
+import prisma from '../db/prismaInstance.js'
+import jwt from 'jsonwebtoken'
+import bcrypt from 'bcryptjs'
+const JWT_SECRET = process.env.JWT_SECRET || 'your-secret-key';
+
+
+
+export const RegisterController = async (req, res) => {
+    try {
+        const { email, username, password, firstName, lastName } = req.body;
+
+        if (!email || !username || !password) {
+        return res.status(400).json({ 
+            error: 'Email, username, and password are required' 
+        });
+        }
+
+        // Check if user exists
+        const existingUser = await prisma.user.findFirst({
+        where: {
+            OR: [{ email }, { username }]
+        }
+        });
+
+        if (existingUser) {
+        return res.status(400).json({ 
+            error: 'User already exists' 
+        });
+        }
+
+        // Hash password
+        const hashedPassword = await bcrypt.hash(password, 10);
+
+        // Create user
+        const user = await prisma.user.create({
+        data: {
+            email,
+            username,
+            password: hashedPassword,
+            firstName,
+            lastName,
+            provider: 'local',
+        }
+        
+        });
+
+        // Generate token
+        const token = jwt.sign({ userId: user.id }, JWT_SECRET, { expiresIn: '24h' });
+
+        res.json({
+        message: 'Registration successful',
+        user: {
+            id: user.id,
+            email: user.email,
+            username: user.username,
+            firstName: user.firstName,
+            lastName: user.lastName,
+            avatar: user.avatar
+        },
+        token
+        });
+
+    } catch (error) {
+        console.error('Registration error:', error);
+        res.status(500).json({ error: 'Registration failed' });
+    }
+};
+
+
+
+export const LoginController = async (req, res) => {
+  try {
+    const { email, password } = req.body;
+
+    if (!email || !password) {
+      return res.status(400).json({ error: 'Email and password are required' });
+    }
+
+    // Find user
+    const user = await prisma.user.findUnique({ where: { email } });
+
+    if (!user) {
+      return res.status(400).json({ error: 'Invalid credentials' });
+    }
+
+    if (!user.password) {
+      return res.status(400).json({ error: 'Login with OAuth provider' });
+    }
+
+    // Check password
+    const isValidPassword = await bcrypt.compare(password, user.password);
+    if (!isValidPassword) {
+      return res.status(400).json({ error: 'Invalid credentials' });
+    }
+
+    // Generate token
+    const token = jwt.sign({ userId: user.id }, JWT_SECRET, { expiresIn: '24h' });
+
+    res.json({
+      message: 'Login successful',
+      user: {
+        id: user.id,
+        email: user.email,
+        username: user.username,
+        firstName: user.firstName,
+        lastName: user.lastName,
+        avatar: user.avatar
+      },
+      token
+    });
+
+  } catch (error) {
+    console.error('Login error:', error);
+    res.status(500).json({ error: 'Login failed' });
+  }
+};
+
+export const ProfileController = async (req, res) => {
+  try {
+    const user = await prisma.user.findUnique({
+      where: { id: req.user.userId },
+      select: {
+        id: true,
+        email: true,
+        username: true,
+        firstName: true,
+        lastName: true,
+        createdAt: true,
+        provider: true,
+        avatar: true
+      }
+    });
+
+    if (!user) {
+      return res.status(404).json({ error: 'User not found' });
+    }
+
+    res.json({ user });
+  } catch (error) {
+    console.error('Profile error:', error);
+    res.status(500).json({ error: 'Profile failed' });
+  }
+};
+
