+++ backend/db/index.js
@@ -1,0 +1,11 @@
+import prisma from "./prismaInstance.js"
+async function connectDB() {
+  try {
+    await prisma.$connect();
+    console.log('Database connected successfully');
+  } catch (error) {
+    console.error('Database connection failed:', error.message);
+  }
+}
+
+export default connectDB;
