+++ backend/db/prismaInstance.js
@@ -1,0 +1,4 @@
+import { PrismaClient } from '@prisma/client'
+
+const prisma = new PrismaClient();
+export default prisma;
