+++ backend/index.js
@@ -1,23 +1,27 @@
-const express = require('express');
-const cors = require('cors');
-const bcrypt = require('bcryptjs');
-const jwt = require('jsonwebtoken');
-const passport = require('passport');
-const GoogleStrategy = require('passport-google-oauth20').Strategy;
-const GithubStrategy = require('passport-github2').Strategy;
-const session = require('express-session');
-const { PrismaClient } = require('@prisma/client');
-require('dotenv').config();
+import express from 'express';
+import cors from 'cors';
+import jwt from 'jsonwebtoken';
+import passport from 'passport';
+import { Strategy as GoogleStrategy } from 'passport-google-oauth20';
+import { Strategy as GithubStrategy } from 'passport-github2';
+import session from 'express-session';
+import connectDB from './db/index.js';
+import prisma from './db/prismaInstance.js';
+import dotenv from 'dotenv';
+import AuthRoutes from './routes/authRoutes.js'
+
+dotenv.config();
+// Connect to database
+connectDB();
 
 const app = express();
-const prisma = new PrismaClient();
 const PORT = process.env.PORT || 5001;
 const JWT_SECRET = process.env.JWT_SECRET || 'your-secret-key';
 
 // Middleware
 app.use(cors({
   origin: 'http://localhost:3000',
-  credentials: true
+  credentials: true  
 }));
 app.use(express.json());
 
@@ -145,16 +149,6 @@ passport.deserializeUser(async (id, done) => {
   }
 });
 
-// Connect to database
-async function connectDB() {
-  try {
-    await prisma.$connect();
-    console.log('Database connected successfully');
-  } catch (error) {
-    console.error('Database connection failed:', error.message);
-  }
-}
-connectDB();
 
 // Root route
 app.get('/', (req, res) => {
@@ -164,196 +158,8 @@ app.get('/', (req, res) => {
   });
 });
 
-// OAuth routes
-app.get('/api/auth/google', passport.authenticate('google', { scope: ['profile', 'email'] }));
-
-app.get('/api/auth/google/callback', 
-  passport.authenticate('google', { failureRedirect: 'http://localhost:3000/login' }),
-  (req, res) => {
-    const token = jwt.sign({ userId: req.user.id }, JWT_SECRET, { expiresIn: '24h' });
-    res.redirect(`http://localhost:3000/dashboard?token=${token}&user=${encodeURIComponent(JSON.stringify({
-      id: req.user.id,
-      email: req.user.email,
-      username: req.user.username,
-      firstName: req.user.firstName,
-      lastName: req.user.lastName,
-      avatar: req.user.avatar
-    }))}`);
-  }
-);
-
-app.get('/api/auth/github', passport.authenticate('github', { scope: ['user:email'] }));
-
-app.get('/api/auth/github/callback', 
-  passport.authenticate('github', { failureRedirect: 'http://localhost:3000/login' }),
-  (req, res) => {
-    const token = jwt.sign({ userId: req.user.id }, JWT_SECRET, { expiresIn: '24h' });
-    res.redirect(`http://localhost:3000/dashboard?token=${token}&user=${encodeURIComponent(JSON.stringify({
-      id: req.user.id,
-      email: req.user.email,
-      username: req.user.username,
-      firstName: req.user.firstName,
-      lastName: req.user.lastName,
-      avatar: req.user.avatar
-    }))}`);
-  }
-);
-
-// Register route
-app.post('/api/auth/register', async (req, res) => {
-  try {
-    const { email, username, password, firstName, lastName } = req.body;
-
-    if (!email || !username || !password) {
-      return res.status(400).json({ 
-        error: 'Email, username, and password are required' 
-      });
-    }
-
-    // Check if user exists
-    const existingUser = await prisma.user.findFirst({
-      where: {
-        OR: [{ email }, { username }]
-      }
-    });
-
-    if (existingUser) {
-      return res.status(400).json({ 
-        error: 'User already exists' 
-      });
-    }
-
-    // Hash password
-    const hashedPassword = await bcrypt.hash(password, 10);
-
-    // Create user
-    const user = await prisma.user.create({
-      data: {
-        email,
-        username,
-        password: hashedPassword,
-        firstName,
-        lastName,
-        provider: 'local',
-      }
-      
-    });
-
-    // Generate token
-    const token = jwt.sign({ userId: user.id }, JWT_SECRET, { expiresIn: '24h' });
-
-    res.json({
-      message: 'Registration successful',
-      user: {
-        id: user.id,
-        email: user.email,
-        username: user.username,
-        firstName: user.firstName,
-        lastName: user.lastName,
-        avatar: user.avatar
-      },
-      token
-    });
-
-  } catch (error) {
-    console.error('Registration error:', error);
-    res.status(500).json({ error: 'Registration failed' });
-  }
-});
-
-// Login route
-app.post('/api/auth/login', async (req, res) => {
-  try {
-    const { email, password } = req.body;
-
-    if (!email || !password) {
-      return res.status(400).json({ error: 'Email and password are required' });
-    }
-
-    // Find user
-    const user = await prisma.user.findUnique({ where: { email } });
-
-    if (!user) {
-      return res.status(400).json({ error: 'Invalid credentials' });
-    }
-
-    if (!user.password) {
-      return res.status(400).json({ error: 'Login with OAuth provider' });
-    }
-
-    // Check password
-    const isValidPassword = await bcrypt.compare(password, user.password);
-    if (!isValidPassword) {
-      return res.status(400).json({ error: 'Invalid credentials' });
-    }
-
-    // Generate token
-    const token = jwt.sign({ userId: user.id }, JWT_SECRET, { expiresIn: '24h' });
-
-    res.json({
-      message: 'Login successful',
-      user: {
-        id: user.id,
-        email: user.email,
-        username: user.username,
-        firstName: user.firstName,
-        lastName: user.lastName,
-        avatar: user.avatar
-      },
-      token
-    });
-
-  } catch (error) {
-    console.error('Login error:', error);
-    res.status(500).json({ error: 'Login failed' });
-  }
-});
-
-// Auth middleware
-const authenticateToken = (req, res, next) => {
-  const authHeader = req.headers['authorization'];
-  const token = authHeader && authHeader.split(' ')[1];
-
-  if (!token) {
-    return res.status(401).json({ error: 'Token required' });
-  }
-
-  jwt.verify(token, JWT_SECRET, (err, user) => {
-    if (err) {
-      return res.status(403).json({ error: 'Invalid token' });
-    }
-    req.user = user;
-    next();
-  });
-};
-
-// Profile route
-app.get('/api/auth/profile', authenticateToken, async (req, res) => {
-  try {
-    const user = await prisma.user.findUnique({
-      where: { id: req.user.userId },
-      select: {
-        id: true,
-        email: true,
-        username: true,
-        firstName: true,
-        lastName: true,
-        createdAt: true,
-        provider: true,
-        avatar: true
-      }
-    });
-
-    if (!user) {
-      return res.status(404).json({ error: 'User not found' });
-    }
-
-    res.json({ user });
-  } catch (error) {
-    console.error('Profile error:', error);
-    res.status(500).json({ error: 'Profile failed' });
-  }
-});
+// routes
+app.use('/api/auth',AuthRoutes);
 
 // Start server
 app.listen(PORT, () => {
