+++ backend/middleware/authMiddleware.js
@@ -1,0 +1,18 @@
+import jwt from 'jsonwebtoken'
+const JWT_SECRET = process.env.JWT_SECRET || 'your-secret-key';
+export const authenticateToken = (req, res, next) => {
+  const authHeader = req.headers['authorization'];
+  const token = authHeader && authHeader.split(' ')[1];
+
+  if (!token) {
+    return res.status(401).json({ error: 'Token required' });
+  }
+
+  jwt.verify(token, JWT_SECRET, (err, user) => {
+    if (err) {
+      return res.status(403).json({ error: 'Invalid token' });
+    }
+    req.user = user;
+    next();
+  });
+};
