+++ frontend/package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "proxy": "http://localhost:5001",
+  "type": "module",
   "dependencies": {
     "@react-oauth/google": "^0.12.2",
     "@testing-library/jest-dom": "^5.16.4",
